// src/app/interfaces/account/account.ts
export interface Account {
    id: string ;
    username: string;
    firstname: string;
    email: string;
    role: string;
    status: string;
    phone: string;
    insertionDate: string;
    blockStatus: string;
    connectionDT: string;
    currentLogin: string;
    blockExpire: string;
    expire: string;
    token: string;
    picture: string;
}